Refactor type utils
