--- src/utils/types.ts
@@ -1,3 +1,0 @@
-export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
-
-export type GetArrayInnerType<T extends readonly any[]> = T[number];
